修复用户信息保存问题

+++ app/src/main/java/com/qihuan/wanandroid/bean/User.kt
@@ -9,7 +9,7 @@ data class User(
     val id: Int,
     val admin: Boolean,
     val chapterTops: List<String>,
-    val collectIds: List<Int>,
+    val collectIds: List<Long>,
     val email: String,
     val icon: String,
     val nickname: String,
